Moved old code to legacy, prepare to complete rework
