+++ src/main/java/dev/syoritohatsuki/duckyupdaterrework/legacy/mixin/MinecraftServerMixin.java
@@ -1,4 +1,4 @@
-package dev.syoritohatsuki.duckyupdaterrework.mixin;
+package dev.syoritohatsuki.duckyupdaterrework.legacy.mixin;
 
 import dev.syoritohatsuki.duckyupdaterrework.DuckyUpdaterReWork;
 import net.minecraft.server.MinecraftServer;
@@ -7,7 +7,7 @@ import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.Inject;
 import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
 
-import static dev.syoritohatsuki.duckyupdaterrework.util.AnsiKt.updateListCliMessage;
+import static dev.syoritohatsuki.duckyupdaterrework.legacy.util.AnsiKt.updateListCliMessage;
 
 @Mixin(MinecraftServer.class)
 public abstract class MinecraftServerMixin {
