+++ src/main/kotlin/dev/syoritohatsuki/duckyupdaterrework/legacy/util/Ansi.kt
@@ -1,4 +1,4 @@
-package dev.syoritohatsuki.duckyupdaterrework.util
+package dev.syoritohatsuki.duckyupdaterrework.legacy.util
 
 import dev.syoritohatsuki.duckyupdaterrework.DuckyUpdaterReWork
 import dev.syoritohatsuki.duckyupdaterrework.DuckyUpdaterReWork.logger
