+++ src/main/resources/duckyupdaterrework.mixins.json
@@ -1,6 +1,6 @@
 {
   "required": true,
-  "package": "dev.syoritohatsuki.duckyupdaterrework.mixin",
+  "package": "dev.syoritohatsuki.duckyupdaterrework.legacy.mixin",
   "compatibilityLevel": "JAVA_17",
   "injectors": {
     "defaultRequire": 1
